gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
redis==5.1.0
docker==7.1.0
psutil==6.0.0
